+++ amazonorders/session.py
@@ -245,9 +245,6 @@ class AmazonSession:
         """
         self.get(self.config.constants.SIGN_OUT_URL)
 
-        if os.path.exists(self.config.cookie_jar_path):
-            os.remove(self.config.cookie_jar_path)
-
         self.session.close()
         self.session = Session()