+++ CHANGELOG.md
@@ -11,8 +11,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 ### Changed
 
 - `AmazonOrders.get_order_history()` now fetches each individual order using `asyncio` for 5-10x performance improvements.
-- Synchronous nature of `AmazonSession`, since this was bad design, and also prevents adding any async support.
 - `AmazonOrders.request` now takes `persist_cookies`, which defaults to `False`. Cookies are now only persisted when the login form is submitted, rather than with every request.
+- `AmazonSession` is now thread-safe.
 - Renamed `constants.Constants.TRANSACTION_HISTORY_LANDING_ROUTE` to `constants.Constants.TRANSACTION_HISTORY_ROUTE`.
 - Renamed `constants.Constants.TRANSACTION_HISTORY_LANDING_URL` to `constants.Constants.TRANSACTION_HISTORY_URL`.
 - Testing environment variables are now consistently prefixed with `AMAZON_`.