+++ CHANGELOG.md
@@ -21,7 +21,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 - `AmazonSession.last_response` and `AmazonSession.last_response_parsed`. These variables forced `AmazonSession` in to a statefulness that means its underlying functionality cannot be made async. Use the response object [AmazonSessionResponse](https://amazon-orders.readthedocs.io/api.html#amazonorders.util.AmazonSessionResponse) instead.
 - `constants.Constants.ORDER_HISTORY_LANDING_URL`.
-- evaluation of `OTP_SECRET_KEY`, use `AMAZON_OTP_SECRET_KEY`.
+- Evaluation of `OTP_SECRET_KEY`, use `AMAZON_OTP_SECRET_KEY`.
 
 ## [3.2.19](https://github.com/alexdlaird/amazon-orders/compare/3.2.18...3.2.19) - 2025-04-21